requirements.txt
streamlit>=1.24.0
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6
pydantic>=2.0.0
pycryptodome==3.20.0
rsa==4.9